shuffle code to avoid circular imports
