+++ src/radical/utils/config.py
@@ -122,7 +122,8 @@ __license__   = "MIT"
 import glob
 import os
 
-from .misc       import find_module, is_string
+from .debug      import find_module
+from .misc       import is_string
 from .misc       import expand_env as ru_expand_env
 from .read_json  import read_json
 from .dict_mixin import dict_merge, DictMixin
