+++ src/radical/utils/testing.py
@@ -7,7 +7,7 @@ __license__   = 'MIT'
 import os
 import sys
 
-from .misc       import import_module
+from .debug      import import_module
 from .read_json  import read_json
